Refactor DE quadrature code (#119)
* add -12 error code for x>=0

* Refactor Hankel DE quadrature code

* Ensure Hankel transform DE functions return int and not double

* add tests for -12 error code

* linting change

+++ include/libhankel.h
@@ -121,9 +121,13 @@ double hankel_transform_DHT(int nu, form_factor_f f, double x, void *f_ctx, doub
  * @param output     pointer to var containing output from transform
  * @param n_eval     integer indicating number of function evaluations (``N_ogata`` in SASfit)
  * @param eps_rel    relative error allowed e.g. 1e-9 (``eps_nriq`` in SASfit)
+ *
+ * @return 0 on success, or a negative status code (see the
+ *         <a href="../usage/status_codes.html">Status Codes</a> page). @p x
+ *         must be finite and greater than zero.
  */
-double hankel_transform_DE_Ooura(int nu, form_factor_f f, double x, void *f_ctx, double *output,
-                                 int n_eval, double eps_rel);
+int hankel_transform_DE_Ooura(int nu, form_factor_f f, double x, void *f_ctx, double *output,
+                              int n_eval, double eps_rel);
 
 /**
  * @brief Computes Hankel transform, using de-quadrature.
@@ -136,9 +140,13 @@ double hankel_transform_DE_Ooura(int nu, form_factor_f f, double x, void *f_ctx,
  * @param output     pointer to var containing output from transform
  * @param n_eval     integer indicating number of function evaluations (``N_ogata`` in SASfit)
  * @param f_max      float indicating starting guess for max in form factor (``h_ogata`` in SASfit)
+ *
+ * @return 0 on success, or a negative status code (see the
+ *         <a href="../usage/status_codes.html">Status Codes</a> page). @p x
+ *         must be finite and greater than zero.
  */
-double hankel_transform_DE_Ogata(int nu, form_factor_f f, double x, void *f_ctx, double *output,
-                                 int n_eval, double f_max);
+int hankel_transform_DE_Ogata(int nu, form_factor_f f, double x, void *f_ctx, double *output,
+                              int n_eval, double f_max);
 
 /**
  * @brief Computes Hankel transform using the Quadrature With Extrapolation method by Key.
